rename swagger.json to openapi.json
gets rid of some warnings as vs sees it as an actual swagger document

also add tenantkeyheaderattribute
