+++ src/Infrastructure/OpenApi/TenantKeyHeaderAttribute.cs
@@ -1,0 +1,15 @@
+﻿using DN.WebApi.Shared.Multitenancy;
+
+namespace DN.WebApi.Infrastructure.OpenApi;
+
+public class TenantKeyHeaderAttribute : SwaggerHeaderAttribute
+{
+    public TenantKeyHeaderAttribute()
+        : base(
+            MultitenancyConstants.TenantHeaderKey,
+            "Input your tenant Id to access this API",
+            string.Empty,
+            true)
+    {
+    }
+}
